Add '-progress_reprots' cmd option to db_bench
Summary:
In db_stress, we have '-progress_reports' that controls whether the
process prints finished ops periodically to stdout. We can have the same
knob for db_bench, which can be handy when redirecting the output of
db_bench to files.

Test Plan:
./db_bench -benchmarks=fillseq -progress_reprots=false
./db_bench -benchmarks=fillseq

The first command disables progress reporting, while the second enables.